requirements.txt
streamlit
cohere
pdfminer.six
ultralytics
opencv-python
python-dotenv

